+++ src/assets/styles/partials/product.scss
@@ -79,7 +79,7 @@ $old-price-color:#8f8f8f;
                 padding-left: 8px !important;
                 padding-right: 10px !important;
                 text-align: left;
-                font-size: 20px;
+                font-size: 16px;
                 font-weight: 100;
                 color: #495057;
             }
